+++ lib/ddtrace/contrib/elasticsearch/quantize.rb
@@ -4,28 +4,26 @@ module Datadog
       # Quantize contains ES-specific resource quantization tools.
       module Quantize
         PLACEHOLDER = '?'.freeze
+        ID_PLACEHOLDER = '\1?'.freeze
         EXCLUDE_KEYS = [].freeze
         SHOW_KEYS = [:_index, :_type, :_id].freeze
-        DEFAULT_OPTIONS = { exclude: EXCLUDE_KEYS, show: SHOW_KEYS }.freeze
-
-        ID_REGEXP = %r{\/([0-9]+)([\/\?]|$)}
-        ID_PLACEHOLDER = '/?\2'.freeze
-
-        INDEX_REGEXP = /[0-9]{2,}/
-        INDEX_PLACEHOLDER = '?'.freeze
+        DEFAULT_OPTIONS = {
+          exclude: EXCLUDE_KEYS,
+          show: SHOW_KEYS,
+          placeholder: PLACEHOLDER
+        }.freeze
 
         module_function
 
-        # Very basic quantization, complex processing should be done in the agent
         def format_url(url)
-          quantized_url = url.gsub(ID_REGEXP, ID_PLACEHOLDER)
-          quantized_url.gsub(INDEX_REGEXP, INDEX_PLACEHOLDER)
+          sanitize_fragment_with_id(url)
+            .gsub(/(?:[\d]+)/, PLACEHOLDER)
         end
 
         def format_body(body, options = {})
           format_body!(body, options)
         rescue StandardError
-          PLACEHOLDER
+          options[:placeholder] || PLACEHOLDER
         end
 
         def format_body!(body, options = {})
@@ -36,46 +34,10 @@ module Datadog
 
           # Parse each statement and quantize them.
           statements.collect do |string|
-            reserialize_json(string) do |obj|
-              format_statement(obj, options)
+            reserialize_json(string, options[:placeholder]) do |obj|
+              Datadog::Quantization::Hash.format(obj, options)
             end
           end.join("\n")
-        end
-
-        def format_statement(statement, options = {})
-          return statement if options[:show] == :all
-
-          case statement
-          when Hash
-            statement.each_with_object({}) do |(key, value), quantized|
-              if options[:show].include?(key.to_sym)
-                quantized[key] = value
-              elsif !options[:exclude].include?(key.to_sym)
-                quantized[key] = format_value(value, options)
-              end
-            end
-          else
-            format_value(statement, options)
-          end
-        end
-
-        def format_value(value, options = {})
-          return value if options[:show] == :all
-
-          case value
-          when Hash
-            format_statement(value, options)
-          when Array
-            # If any are objects, format them.
-            if value.any? { |v| v.class <= Hash || v.class <= Array }
-              value.collect { |i| format_value(i, options) }
-            # Otherwise short-circuit and return single placeholder
-            else
-              PLACEHOLDER
-            end
-          else
-            PLACEHOLDER
-          end
         end
 
         def merge_options(original, additional)
@@ -104,6 +66,13 @@ module Datadog
             # If it can't parse/dump, don't raise an error.
             fail_value
           end
+        end
+
+        # Sanitizes URL fragment by changing it to ? whenever a number is detected
+        # This is meant as simple heuristic that attempts to detect if particular fragment
+        # represents document Id. This is meant to reduce the cardinality in most frequent cases.
+        def sanitize_fragment_with_id(url)
+          url.gsub(%r{^(/?[^/]*/[^/]*/)(?:[^\?/\d]*[\d]+[^\?/]*)}, ID_PLACEHOLDER)
         end
       end
     end