+++ lib/ddtrace/contrib/mongodb/subscribers.rb
@@ -18,14 +18,15 @@ module Datadog
           Thread.current[:datadog_mongo_span] = span
 
           # build a quantized Query using the Parser module
-          query = Datadog::Contrib::MongoDB.query_builder(event.command_name, event.database_name, event.command)
+          query = Datadog::Contrib::MongoDB
+                  .query_builder(event.command_name, event.database_name, event.command)
           serialized_query = query.to_s
 
           # add operation tags; the full query is stored and used as a resource,
           # since it has been quantized and reduced
-          span.set_tag(Datadog::Ext::Mongo::DB, query[:database])
-          span.set_tag(Datadog::Ext::Mongo::COLLECTION, query[:collection])
-          span.set_tag(Datadog::Ext::Mongo::OPERATION, query[:operation])
+          span.set_tag(Datadog::Ext::Mongo::DB, query['database'])
+          span.set_tag(Datadog::Ext::Mongo::COLLECTION, query['collection'])
+          span.set_tag(Datadog::Ext::Mongo::OPERATION, query['operation'])
           span.set_tag(Datadog::Ext::Mongo::QUERY, serialized_query)
           span.set_tag(Datadog::Ext::NET::TARGET_HOST, event.address.host)
           span.set_tag(Datadog::Ext::NET::TARGET_PORT, event.address.port)